Add files via upload
#<h1>–<h6>: The HTML Section Heading elements#
The <h1> to <h6> HTML elements represent six levels of section headings. 
<h1> is the highest section level and <h6> is the lowest. 
By default, all heading elements create a block-level box in the layout, starting on a new line and taking up the full width available in their containing block.

#<p>: Paragraph tag
In HTML, the <p> tag is used to define paragraphs of text. It's a fundamental element for structuring the content of a web page. 
This tag is essential for organizing and formatting text on a webpage.

#<br>: Break line tag
In HTML, the <br> tag is used to create a line break or a line of white space within your content. 
It is an empty, self-closing tag, which means it doesn't have a closing tag like some other HTML elements. 
When you use <br>, it simply moves the content that follows it to the next line.